requirements.txt
websocket-client
requests
python-dotenv
pandas
streamlit
